refactor: HealthSystem health to int

+++ Assets/Project/Scripts/Interface/IHealth.cs
@@ -2,15 +2,15 @@ using System;
 
 public interface IHealth
 {
-    float CurrentHealth { get; }
-    float MaxHealth { get; }
+    int CurrentHealth { get; }
+    int MaxHealth { get; }
     bool IsDead { get; }
 
     event Action OnDeath;
-    event Action<float, float> OnHealthChanged;
+    event Action<int, int> OnHealthChanged;
 
     void Initialize(int maxHealth);
-    void TakeDamage(float damage);
-    void Heal(float amount);
-    void SetMaxHealth(float newMax);
+    void TakeDamage(int damage);
+    void Heal(int amount);
+    void SetMaxHealth(int newMax);
 }
